消除警告

+++ TrafficMonitor/Common.h
@@ -19,7 +19,7 @@ public:
     {
         if (str.empty()) return;
 
-        int size = str.size();  //字符串的长度
+        int size = static_cast<int>(str.size());  //字符串的长度
         if (size < 0) return;
         int index1 = 0;     //字符串中第1个不是空格或控制字符的位置
         int index2 = size - 1;  //字符串中最后一个不是空格或控制字符的位置
@@ -257,8 +257,8 @@ public:
     template<class T>
     static double StringSimilarDegree_LD(const T& srcString, const T& matchString)
     {
-        int n = srcString.size();
-        int m = matchString.size();
+        int n = static_cast<int>(srcString.size());
+        int m = static_cast<int>(matchString.size());
         //int[, ] d = new int[n + 1, m + 1]; // matrix
         vector<vector<int>> d(n + 1, vector<int>(m + 1));
         int cost; // cost
